+++ plugins/css-blank-pseudo/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to CSS Blank Pseudo
 
+### Unreleased
+
+- Fix: assign global browser polyfill to `window`, `self` or a blank object.
+
 ### 4.1.0 (July 30, 2022)
 
 - Added: `disablePolyfillReadyClass` plugin option to prevent `.js-blank-pseudo` from being added.