+++ plugins/css-has-pseudo/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to CSS Has Pseudo
 
+### Unreleased
+
+- Fix: assign global browser polyfill to `window`, `self` or a blank object.
+
 ### 4.0.0 (July 8, 2022)
 
 [Read the full changelog](https://github.com/csstools/postcss-plugins/wiki/PostCSS-Preset-Env-8)